Revert "change some uint64_t to uint32_t"
This reverts commit 8e328696128c9c5f736e57300b6cfc780e37ea2f.

+++ ComProxy_ut.cpp
@@ -27,7 +27,7 @@
 
 #define CRC_SIZE 2
 
-ClientSocket::ClientSocket(uint32_t addr, uint16_t port, int32_t style) : mSock(0) {}
+ClientSocket::ClientSocket(uint64_t addr, uint16_t port, int32_t style) : mSock(0) {}
 ClientSocket::~ClientSocket(void) {}
 
 const BlInfo dummyBlInfo = {0xDE, 0xAD, 0xAF, 0xFE, 0xFF, 0x4, 0x0, 0xB0, 0xB1, 0xE5, 0x00};
@@ -45,7 +45,7 @@ size_t g_TestSocketRecvBufferPos = 0;
 size_t g_TestSocketRecvBufferSize = 0;
 uint8_t g_TestSocketSendBuffer[10240];
 size_t g_TestSocketSendBufferSize;
-TestSocket::TestSocket(uint32_t	addr, uint16_t port) : ClientSocket(addr, port, 0)
+TestSocket::TestSocket(uint64_t	addr, uint16_t port) : ClientSocket(addr, port, 0)
 {
 	m_Delay.tv_sec = 0;
 	m_Delay.tv_nsec = 0;
